tools/atrf-xmit/atrf-xmit.c: corrected title comment

+++ tools/atrf-proxy/atrf-proxy.c
@@ -169,7 +169,7 @@ static int cmd_more(struct atrf_dsc *dsc, struct netio *netio, const char *cmd)
 		return netio_printf(netio, "+0x%02x\n", res);
 	}
 	if (!strcasecmp(cmd, "wait")) {
-		uint8_t res;
+		int res;
 
 		if (!n)
 			n = 1;
